+++ src/main/java/com/iglusoft/api/database/entities/Offer.java
@@ -29,7 +29,7 @@ public class Offer {
     private DiscountType discountType;
 
     @Positive
-    private Long discountAmount;
+    private Double discountAmount;
 
     public Offer() {
     }
@@ -66,11 +66,11 @@ public class Offer {
         this.discountType = discountType;
     }
 
-    public Long getDiscountAmount() {
+    public Double getDiscountAmount() {
         return discountAmount;
     }
 
-    public void setDiscountAmount(Long discountAmount) {
+    public void setDiscountAmount(Double discountAmount) {
         this.discountAmount = discountAmount;
     }
 }
